+++ app/core/tests/unit/ticket_category/test_unit_ticket_category_serializer.py
@@ -1,0 +1,26 @@
+import pytest
+
+from api.tests.unit.test_unit_serializer import (
+    SerializerTestCases
+)
+
+
+
+@pytest.mark.model_ticketcategory
+class TicketCategorySerializerTestCases(
+    SerializerTestCases
+):
+    pass
+
+
+class TicketCategorySerializerInheritedCases(
+    TicketCategorySerializerTestCases
+):
+    pass
+
+
+@pytest.mark.module_core
+class TicketCategorySerializerPyTest(
+    TicketCategorySerializerTestCases
+):
+    pass