+++ Release-Notes.md
@@ -1,3 +1,12 @@
+## Version v1.28.0
+
+- Change Ticket model added behind feature flag `2025-00009` and will remain behind this flag until production ready.
+
+- Incident Ticket model added behind feature flag `2025-00010` and will remain behind this flag until production ready.
+
+- Problem Ticket model added behind feature flag `2025-00011` and will remain behind this flag until production ready.
+
+
 ## Version v1.27.0
 
 - Model [Company](https://nofusscomputing.com/projects/centurion_erp/user/access/company/) now available, as feature flag `2025-00008` has been removed from Centurion ERPs feature flag checks.